streaming: Fix bug in `stream_queryset_as_xlsx()` when used with Django QuerySet objects
183d794aee94e1127dd20c75a (version 0.4.0) accidentally changed the
behaviour of this function in some circumstances. When given a Django
QuerySet, the queryset would be evaluated only once (thus making a
single SQL query), and batched manually by the function. This could
cause performance issues if there were many rows to fetch. The
previous behaviour (which we restore here) was to perform multiple SQL
queries, actually using the `batch_size` argument of the function.

This bug was spotted by Vincent Desprez (@vinzd).